feat: initialize config more flexibly
APIConfig can now be initialized from a CLI Flag pointing to a
json file. If this flag is omitted, the CUSTOM_CONFIG env var
will be sourced for the json string with the config values.

Additionally, any environment variables wtih the prefix
PIXLISE_CONFIG_* will be added to the APIConfig struct
or will be used to override any present values in config