Updated Image GET endpoint to check for: - Image in DB - User permission to the image (using scan IDs from DB entry associatedScanIds array), using membership cache (meaning if user is not recently connected to web socket, it will fail, which should be a good thing?!) - Modifier query params (thumbnail and with-location supported so far)
And it also generates the modified image and puts it in dataset bucket Image-Cache dir for future requests to find

As part of this, a few places where we check for ownership of scan related data have been simplified, saving a DB read.

Also fixed bug with first websocket connection of new user (if user not in DB yet) - it was not handling it the same as a regular user connection, so things like auth0| prepend to userid and group lookup didnt happen, now it's uniform.

Removed auth0| snip from incoming JWT user ID, so all endpoint handlers get the same user id as the web socket code now

Integration test now has access to the API's configured dataset bucket and can seed/delete files as needed for test

Testing of Image GET is done in 2 steps - pre websocket tests, and tests executed as part of testing scan data

This also resulted in a new core package imageedit which has generic functions for getting bytes from an image (encoded as JPG or PNG), image scaling, and ability to mark points on an image.